Change the gcc options for sdl builds to allow for gnu99 features, it needs some fixes in other places. Fixes test_mem compilation failure on cygwin.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@28723 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/test_disk.c
@@ -434,7 +434,7 @@ enum plugin_status plugin_start(const void* parameter)
 
     audiobuf = rb->plugin_get_audio_buffer(&audiobuflen);
     /* align start and length to 32 bit */
-    align = (-(int)audiobuf) & 3;
+    align = (-(intptr_t)audiobuf) & 3;
     audiobuf += align;
     audiobuflen = (audiobuflen - align) & ~3;
